online_store_on_sofa_project/requirements.txt
asgiref==3.3.0
certifi==2020.11.8
chardet==3.0.4
Django==3.1.3
django-crispy-forms==1.9.2
django-discover-runner==1.0
django-g-recaptcha==0.2.4
django-ranged-response==0.2.0
django-role-permissions==3.1.1
django-simple-captcha==0.5.13
funcsigs==1.0.2
idna==2.10
mock==4.0.2
numpy==1.20.0
pandas==1.2.1
pbr==5.5.1
Pillow==8.0.1
python-dateutil==2.8.1
pytz==2020.4
reportlab==3.5.59
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
stripe==2.55.1
urllib3==1.25.11
